rename demography store
