+++ app/src/app/utils/map/mapRenderSubs.ts
@@ -8,7 +8,7 @@ import {
 } from '@utils/helpers';
 import {useMapStore as _useMapStore, MapStore} from '@store/mapStore';
 import {getFeatureUnderCursor} from '@utils/helpers';
-import {useDemographyStore as _useDemographyStore} from '../../store/demographicMap';
+import {useDemographyStore as _useDemographyStore} from '../../store/demographyStore';
 import {useHoverStore as _useHoverStore, HoverFeatureStore} from '../../store/hoverFeatures';
 import {demographyCache} from '../demography/demographyCache';
 import {FocusState, ShatterState} from './types';
